requirements.txt
pysnmp==7.1.20
paho-mqtt
pyasn1
pysmi